+++ examples/echo_service/prod_test/gen_cfgs/prod_2.go
@@ -1,0 +1,12 @@
+package main
+
+const (
+	cNetworkKey_2 = "oi4r9NW9Le7fKF9d"
+)
+
+var (
+	gListOfConnects_2 = []string{
+		"193.233.18.245:9581", // 1x2.5GHz, 1.0GB RAM, 5GB VNMe
+		"62.233.46.109:9581",  // 1x2.8GHz, 1.0GB RAM, 16GB SSD
+	}
+)